+++ quilt/changelog.md
@@ -1,1 +1,1 @@
-- Updated to 26.1
+- Added access to the player in the voice host event